This is a proof of concept for how to integrate recurring events
